+++ cla/bbc/connector_test.go
@@ -27,7 +27,7 @@ func TestConnector(t *testing.T) {
 		t.Fatal(bErr)
 	}
 
-	if err := c.Send(&b); err != nil {
+	if err := c.Send(b); err != nil {
 		t.Fatal(err)
 	}
 
@@ -51,7 +51,7 @@ func TestConnectorUnregisterTransmission(t *testing.T) {
 		t.Fatal(bErr)
 	}
 
-	if err := c.Send(&b); err != nil && err.Error() != "peer send failure Fragment" {
+	if err := c.Send(b); err != nil && err.Error() != "peer send failure Fragment" {
 		t.Fatal(err)
 	}